+++ judge-impl/src/main/java/com/dwarfeng/judge/impl/service/EvaluateQosServiceImpl.java
@@ -1,5 +1,6 @@
 package com.dwarfeng.judge.impl.service;
 
+import com.dwarfeng.judge.stack.bean.ConsumerStatus;
 import com.dwarfeng.judge.stack.bean.EvaluateInfo;
 import com.dwarfeng.judge.stack.handler.ConsumeHandler;
 import com.dwarfeng.judge.stack.handler.EvaluateHandler;
@@ -17,6 +18,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import javax.annotation.PreDestroy;
+import java.util.Objects;
 import java.util.concurrent.locks.Lock;
 import java.util.concurrent.locks.ReentrantLock;
 
@@ -106,6 +108,46 @@ public class EvaluateQosServiceImpl implements EvaluateQosService {
             evaluateLocalCacheHandler.clear();
         } catch (Exception e) {
             throw ServiceExceptionHelper.logAndThrow("清除本地缓存时发生异常",
+                    LogLevel.WARN, sem, e
+            );
+        } finally {
+            lock.unlock();
+        }
+    }
+
+    @SuppressWarnings("DuplicatedCode")
+    @Override
+    public ConsumerStatus getConsumerStatus() throws ServiceException {
+        lock.lock();
+        try {
+            return new ConsumerStatus(
+                    consumeHandler.bufferedSize(),
+                    consumeHandler.getBufferSize(),
+                    consumeHandler.getThread(),
+                    consumeHandler.isIdle()
+            );
+        } catch (Exception e) {
+            throw ServiceExceptionHelper.logAndThrow("获取消费者状态时发生异常",
+                    LogLevel.WARN, sem, e
+            );
+        } finally {
+            lock.unlock();
+        }
+    }
+
+    @SuppressWarnings("DuplicatedCode")
+    @Override
+    public void setConsumerParameters(Integer bufferSize, Integer thread) throws ServiceException {
+        lock.lock();
+        try {
+            consumeHandler.setBufferSize(
+                    Objects.isNull(bufferSize) ? consumeHandler.getBufferSize() : bufferSize
+            );
+            consumeHandler.setThread(
+                    Objects.isNull(thread) ? consumeHandler.getThread() : thread
+            );
+        } catch (Exception e) {
+            throw ServiceExceptionHelper.logAndThrow("设置消费者参数时发生异常",
                     LogLevel.WARN, sem, e
             );
         } finally {